+++ src/composer/compose_predicate.rs
@@ -4,7 +4,6 @@ use crate::ast::dependent_function::DependentFunction;
 use crate::ast::function_type::FunctionType;
 use crate::ast::binary_op::BinOperator;
 use std::collections::HashMap;
-use crate::composer::postulate::{DefinitionInserter, PostulateInserter};
 use crate::monty::fresh_variable::to_unicode_subscript;
 use crate::{app, bin_op, dependent_function, function_type, record_projection, term, var_decl};
 use crate::ast::agda_expr::AgdaExpr;
@@ -12,7 +11,7 @@ use crate::ast::agda_expr::AgdaExpr::{UnOp};
 use crate::ast::agda_expr::AgdaExpr::Term;
 use crate::ast::application::Application;
 use crate::ast::operator::Operator::{Necessity, Product};
-use crate::ast::program::Program;
+use crate::ast::program::{DefinitionInserter, PostulateInserter, Program};
 use crate::ast::record_decl::Record;
 use crate::ast::top_decl::TDeclaration::RecordDecl;
 use crate::ast::unary_op::UnOperator;
