+++ src/composer/compose_variable.rs
@@ -1,14 +1,10 @@
 use crate::ast::record_projection::RecordProjection;
 use crate::ast::function_type::FunctionType;
-use crate::composer::case_converter::{convert_case, CaseStyle};
-use crate::composer::postulate::{DefinitionInserter, PostulateInserter};
-use crate::{app, function_type, record_projection, term};
 use crate::ast::agda_expr::AgdaExpr;
 use crate::ast::agda_expr::AgdaExpr::Term;
-use crate::ast::agda_expr::AgdaExpr::App;
+use crate::ast::top_decl::TDeclaration;
 use crate::ast::application::Application;
-use crate::ast::postulate_decl::Postulate;
-use crate::ast::program::Program;
+use crate::ast::program::{DefinitionInserter, PostulateInserter, Program};
 use crate::ast::record_decl::Record;
 use crate::ast::top_decl::TDeclaration::RecordDecl;
 use crate::ast::var_declaration::VarDecl;
@@ -16,6 +12,8 @@ use crate::composer::compose_predicate::generate_predicate_output;
 use crate::composer::lambda_to_types::generate_function_header;
 use crate::composer::langtree::Token;
 use crate::monty::fresh_variable::to_unicode_subscript;
+use crate::composer::case_converter::{convert_case, CaseStyle};
+use crate::{app, function_type, record, record_projection, term, var_decl};
 
 pub fn compose_variable(token: Token, f: &mut Program, props: Vec<Token>) -> (String, AgdaExpr) {
 
@@ -86,21 +84,13 @@ pub fn compose_variable(token: Token, f: &mut Program, props: Vec<Token>) -> (St
     let record_name = format!("{}ᵣ", convert_case(props_iden.clone().as_str(), CaseStyle::PascalCase));
     let constructor_name = format!("{}꜀", convert_case(props_iden.clone().as_str(), CaseStyle::PascalCase));
 
-    let rec = Record {
-        record_iden: record_name.clone(),
-        constructor_iden: constructor_name,
-        fields: fields,
-        comment : None
-    };
+    let rec = record!(record_name.clone(), constructor_name, fields, None);
 
-    let postulate_entry = VarDecl {
-        iden: predicate_iden.clone(),
-        _type: Box::from(generate_function_header(1)),
-    };
+    let postulate_entry = var_decl!(predicate_iden.clone(), generate_function_header(1));
 
     /* We need to also update the postulate to include the isType function */
     f.insert_postulate(postulate_entry);
-    f.insert_definition(RecordDecl(rec));
+    f.insert_definition(rec);
 
     let proj = record_projection!(*term!(record_name.clone()), *term!("e₁"));
     let projection = app!(proj, *term!("e₁"));
