+++ src/composer/mod.rs
@@ -1,16 +1,11 @@
 
 pub mod lambda_to_types;
 pub mod agdaify;
-pub mod postulate;
-mod record;
 pub mod case_converter;
 pub mod knowledge_base;
 pub mod conclusions;
-pub mod function_def;
 mod compose_predicate;
 mod compose_variable;
 pub mod langtree;
-mod ast;
-mod ast_format;
 mod synonym_handler;
 mod structures;
